Add files via upload
Added following functionality:
1) View water flow in real-time
2) Read total flow from terminal
3) Adjustable configuration data stored in ROM (site #, ID #, meter size)
4) Configuration data stored in on-chip EEPROM
5) Fixed overflow issue in displayed total flow (8-bit number to 32-bit number)

+++ src/piggy_back_loggers/SD_Mag_LIS3MDL_Prototype/Firmware/Firmware.ino
@@ -1,7 +1,7 @@
 // Firmware for the CIWS Residential Datalogger
 // Arduino IDE ver. 1.8.8
 // Utah Water Research Lab
-// Updated: 5/7/2019 (changed to be compatible with Pololu LIS3MDL sensor board)
+// Updated: 5/8/2019 (changed to be compatible with Pololu LIS3MDL sensor board)
 // Daniel Henshaw and Josh Tracy
 // Note: F("String") keeps string literals in program memory and out of RAM. Saves RAM space. Very good. Don't remove the F. I know it looks funny. But don't do it. Really. The program might crash if you do. And then you'll have dishonor on yourself, dishonor on your cow, and you'll find out your cricket ain't lucky.
 // Note: Be sure to process resulting data file as ASCII characters, not Unicode. 
@@ -323,7 +323,7 @@ void storeNewRecord()
     finalCount = State.pulseCount;                      //     Store pulse count to a variable named final count
     State.pulseCount = 0;                               //     Set pulseCount to zero
     State.lastCount = finalCount;
-    State.totalCount += finalCount;
+    State.totalCount += (unsigned int)finalCount;
                                                         //     Read current time from the Real Time Clock and update the Date struct with the current time                                                      
       temp         = rtcTransfer(reg_Years,  READ, 0);  //         read the Years and store into temp variable
       Date.years   = bcdtobin(temp, YEARS_REG_MASK);    //         convert from binary-coded decimal into binary, and store into years field of Date struct
