+++ src/piggy_back_loggers/SD_Mag_LIS3MDL_Prototype/Firmware/state.h
@@ -9,7 +9,7 @@ typedef struct
 {
   volatile byte pulseCount;
   volatile byte lastCount;
-  volatile byte totalCount;
+  volatile unsigned int totalCount;
   volatile unsigned long recordNum;
   volatile bool logging;
   volatile bool flag4;
